disallow ai.approvals_agent with loops
